Fix pixel format

+++ src/pixel.rs
@@ -6,10 +6,10 @@ pub struct Pixel {
     pub x: f32,
     pub y: f32,
     pub z: f32,
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
-    pub a: u8,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
+    pub a: f32,
     pub size_factor: f32,
 }
